Extract tool-call execution into _tool_execution.py
Move process_tool_calls and its supporting functions/classes out of the
2.7k-line _agent_graph.py into a dedicated module. Pure code move, no behavior
change; _agent_graph re-imports process_tool_calls so existing call sites
(CallToolsNode, Agent.run_stream) are unaffected.
